+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Fix for `s3_file_download` to allow multiple file paths passed to `new_path` (#34), thanks to @sckott for contribution.
 * Fix for `s3_file_info`, convert `logical(0)` to `NA` to correctly build `data.frame` output.
+* improve helper function `str_split` performance
 
 # s3fs 0.1.4